Try to get the DCS bug in 1.5.3 resolved with :isExist for a Client Group
Client groups connected remotely to a server, are not generating any
birth events on the server. The server therefore thinks that the Group
does not exist ...

This bug is present since 1.5.

Currently building a logic to go "around" that problem using other
functions (the coalition:getPlayerUnits() is key for resolvement)