Optimization of Cargo Loading behaviour
Cargo loading is a bit complex.
- Cargo should only respawn if it is not existing, and if it is not
loading to another client.
- The pickup menu should allow Cargo that was loading (but for some
reason it did not load), to retry the loading process, but only when the
Client is the Client of the active player.
- When the menus are removed, any unit that was loading but not loaded,
should become back to status none, but only when the Client is the
Client of the active player.
- Fixed a couple of errors with the Pickup-menu function.