Fix bug when 2 humans are in the same group of planes
If more than one human is in a plane's group, the menu bug and a lot of PSEUDOATC menu are shown.
The new menu structure is:
- F10
-  PSEUDOATC
-   <PlayerName>
-    <PSEUDOATC Functions>

For achieving that goal, the self structure has changed from self.player[GID] to self.group[GID].player[UID]. Doing that way, if the last player pit leave the group this functionality do not generate a nil error anymore.
Some text have been changed to indicate CallSign or PlayerName.