Urgent fix for Ford_AB name change. this breaks Moose completely until pushed.
ATC table was changed to the correct name for the enumerator string of "Ford" to "Ford_AB" The main enumerator table was not also changed so the two mismatched.
One has to agree with the other else Moose stops with an index error.
However, there could be minor risk if someone refers to the old enumerator. Whilst we dont care about the string name, the AB was changed by ED anyway. The actual way to keep is to have the string equal the map name. Either way is damned if you do, damned if you don't but I'd rather keep the convention and blame ED than keep it wrong and inconsistent.