Update CTLD.lua
Issue was, 
When you had a Engineer as cargo, it would display it in the troops section, it would not show as cargo.
after I fixed that it opened the door for more issues, now the troops engineer was showing inside the cargo. 

long story short, it should be fixed.

Added a way to differentiate between troops engineer and cargo using CanLoadDirectly and LoadDirectly

also, it seemed that when I dropped ALL the cargo, it did not remove it from the _ListCargo.

Sanity check would be good. all though I have tested it.